+++ test/formatters/installers/fourmolu.bash
@@ -1,0 +1,4 @@
+apt-get install -y haskell-platform
+cabal update
+cabal install fourmolu
+cp "${HOME}/.cabal/bin/fourmolu" /usr/local/bin